+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Stringify: Improve stringification performance for inline arrays and tables that fit on one line ([#314])
+
 ## [3.1.2] - 2026-09-22
 
 ### Changed
@@ -454,5 +458,6 @@ This first forked version from [timhall/toml-patch](https://github.com/timhall/t
 [#311]: https://github.com/DecimalTurn/toml-patch/pull/311
 [#312]: https://github.com/DecimalTurn/toml-patch/pull/312
 [#313]: https://github.com/DecimalTurn/toml-patch/pull/313
+[#314]: https://github.com/DecimalTurn/toml-patch/pull/314
 [0e66e68]: https://github.com/DecimalTurn/toml-patch/commit/0e66e68cbf42a07bc44445e46c3ea7bea97f95c1
 [f97d571]: https://github.com/DecimalTurn/toml-patch/commit/f97d57194952b6603f0fcd686c32d3ae52a2c903